% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_question.r
\name{add_question}
\alias{add_question}
\title{Add a question to a dataframe.}
\usage{
add_question(
  question,
  image = NA,
  explanation = NA,
  correct = NA,
  incorrect = NA,
  keywords = NA,
  df = NA
)
}
\arguments{
\item{question}{Question text}

\item{image}{Image filename (optional)}

\item{explanation}{Explanation for the correct answer (optional)}

\item{correct}{Vector of possible correct answers}

\item{incorrect}{Vector of possible incorrect answers}

\item{keywords}{Vector of strings to insert into the question text}

\item{df}{Dataframe to add question to}
}
\value{
The provided dataframe \code{df}, with an additional row. If a dataframe
is not provided, a new one is created.
}
\description{
Add a multiple-choice or single-choice question to a dataframe,
validating and/or transforming inputs.
}
\section{Credits}{
Brighton Pauli, 2022.
}

\examples{
df <- examling::build_question_df()
df <- add_question(
 question = "What color is a male cardinal?",
 correct = c("Red"),
 incorrect = c("Green", "Blue"),
 df = df
)
df

}
