% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_string_question.r
\name{add_string_question}
\alias{add_string_question}
\title{Add a string-type question to a dataframe compatible with schoice, etc.}
\usage{
add_string_question(
  question,
  image = NA,
  explanation = NA,
  correct = NA,
  keywords = NA,
  df = NA
)
}
\arguments{
\item{question}{Question text.}

\item{image}{Image filename (optional)}

\item{explanation}{Explanation for the correct answer (optional).}

\item{correct}{Correct answer, string.}

\item{keywords}{Vector of strings to insert into the question text.}

\item{df}{Dataframe to add question to.}
}
\description{
Add a string-type question to a dataframe compatible with schoice, etc.
}
\section{Credits}{
Brighton Pauli, 2023
}

\seealso{
Wrapper for \code{add_question()}.
}
