% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_question_df.R
\name{build_question_df}
\alias{build_question_df}
\title{Build a dataframe for questions.}
\usage{
build_question_df()
}
\description{
Build a question dataframe, with question, image, correct,
incorrect, and explanation columns.
}
\section{Credits}{
Brighton Pauli, 2022.
}

