% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulleted_list.r
\name{bulleted_list}
\alias{bulleted_list}
\title{Concatenate string items to form a bulleted list.}
\usage{
bulleted_list(x)
}
\arguments{
\item{x}{vector of strings}
}
\description{
Concatenate string items to form a bulleted list.
}
\section{Credits}{
Brighton Pauli, 2022.
}

\examples{
bulleted_list(c("a", "b", "c"))
}
