% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_sections.R
\name{comment_out_sections_}
\alias{comment_out_sections_}
\title{Help remove all sections from a pre-created QTI zip file.}
\usage{
comment_out_sections_(qti_file, temp_folder, out_file = NULL)
}
\description{
Help remove all sections from a pre-created QTI zip file.
}
\section{Credits}{
Brighton Pauli, 2022
}

