% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmd.R
\name{create_rmd}
\alias{create_rmd}
\title{Start a new .Rmd file from a template.}
\usage{
create_rmd(save_as, question_type, dynamic_type = "none", open = false)
}
\arguments{
\item{save_as}{Where to save the created file.}

\item{question_type}{What kind of question being created. From:
\itemize{
\item \code{"schoice"}
\item \code{"mchoice"}
\item \code{"string"}
}}

\item{dynamic_type}{How to make the question dynamic. Default is \code{"none"}.
From:
\itemize{
\item \code{"none"} - No code blocks are used for random question generation.
Normal.
\item \code{"variations"} - Create a dataframe that contains full variations of
a question.
\item \code{"keywords"} - Like \code{"variations"}, but only a few key words in the
question are changed between variations, as well as the answers.
\item \code{"dynamic"} - (\code{"schoice"} and \code{"mchoice"} only). Start with a tribble of
potential answers and draw from it for each variation. Other answers
from the tribble are used as incorrect choices.
\item \code{"pooled"} - (\code{"schoice"} and \code{"mchoice"} only). Single question where
the answer choices are randomly selected.
}}

\item{open}{Whether or not to open the file upon creation. Default: \code{false}.}
}
\description{
Create a .Rmd file in a specified location with the basic
layout provided, ready for the user to create a new question.
}
\details{
Examples for Dynamic Types
\itemize{
\item \code{"none"} - Q: "Who developed the famous equation $e = mc^2$?", A: "Albert
Einstein".
\item \code{"variations"} - Different questions referring to the properties of
helium atoms that should still be kept in the same file. Q:
"What is the atomic mass of helium?", A: "4". Q: "What state of
matter is helium", A: "Gas".
\item \code{"keywords"} - Questions asking the atomic masses of different elements.
Q: "What is the atomic mass of HELIUM?", A: "4". Q: "What is the
atomic mass of OXYGEN?", A: "16".
\item \code{"dynamic"} - Questions asking which elements belong to different columns.
Q: "Which element(s) can be found in column 18?", A: "He", "Ne", "Ar",
"Kr", "Xe", "rn", or "Og". Q: "Which element(s) can be found in column
16?", A: "O", "S", "Se", "Te", "Po", or "Lv".
\item \code{"pooled"} - Q: "Which element has the largest atomic radius?",
A: "Francium" (the incorrect answers can be drawn from a pool
containing all other elements).
}
}
\section{\code{string} Questions}{
If the \code{question_type} is \code{"string"}, the \code{dynamic_type} can only be
\code{"none"}, \code{"keywords"}, or \code{"variations"}.
}

\section{Credits}{
Brighton Pauli, 2023
}

