% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2rmd.r
\name{csv2rmd}
\alias{csv2rmd}
\title{Convert a csv file of questions into r/exams-style Rmd files.}
\usage{
csv2rmd(x, output_dir, ...)
}
\arguments{
\item{x}{Filepath to csv file.}

\item{output_dir}{Directory to write Rmd files to.}

\item{...}{Additional arguments for read.csv.}
}
\description{
Convert a csv file of questions into r/exams-style Rmd files.
}
\section{csv format}{
The csv file should have the following columns:
\itemize{
\item Question: the question text
\item Image: the image file's name
\item Ans1: the answer text for the first answer
\item Ans2: the answer text for the second answer
\item Ans3: the answer text for the third answer
\item Ans4: the answer text for the fourth answer
\item Ans5: the answer text for the fifth answer
\item Correct: the correct answer number (1-5)
\item Category: the category of the question
\item SubCat: the subcategory of the question
}
}

\section{Credits}{
Brighton Pauli, 2022.
}

\seealso{
\link{xlsx2rmd}
}
