% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rmd.r
\name{df2rmd}
\alias{df2rmd}
\title{Import questions from an R dataframe.}
\usage{
df2rmd(df, output_dir)
}
\arguments{
\item{df}{Dataframe of questions.}

\item{output_dir}{Directory to write Rmd files to.}
}
\description{
Convert a dataframe of questions into r/exams-style Rmd files.
}
\section{Credits}{
Inspired by Achim's csv2rmd function.
Brighton Pauli, 2022.
}

