% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_sections.R
\name{drop_sections}
\alias{drop_sections}
\title{Remove all sections from a pre-created QTI zip file.}
\usage{
drop_sections(qti_file, out_file = NULL)
}
\arguments{
\item{qti_file}{.zip file to edit}

\item{out_file}{.zip file to write to.}
}
\description{
Remove all sections from a pre-created QTI zip file.
}
\details{
If \code{out_file} is not given, the default is the same as \code{qti_file}, but
with \verb{_nosect} appended before the extension.

If \code{out_file} is given, it must end with the extension \code{.zip}.
}
\section{Credits}{
Brighton Pauli, 2022
}

