% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_sections.r
\name{drop_sections_helper}
\alias{drop_sections_helper}
\title{Help remove all sections from a pre-created QTI zip file.}
\usage{
drop_sections_helper(qti_file, temp_folder, out_file = NULL)
}
\description{
Help remove all sections from a pre-created QTI zip file.
}
\section{Credits}{
Brighton Pauli, 2022
}

