% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_func.r
\name{end_logs}
\alias{end_logs}
\title{Wrapper for \code{logr::log_close}.}
\usage{
end_logs()
}
\description{
Run this function to stop generating logs for your functions.
You can restart the logs using \code{start_logs()}.
}
