% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2js.r
\name{exams2js}
\alias{exams2js}
\title{Create a simple Javascript-powered quiz for testing purposes.}
\usage{
exams2js(
  file,
  n = 1L,
  nsamp = NULL,
  dir = ".",
  template = NULL,
  name = NULL,
  quiet = TRUE,
  edir = NULL,
  tdir = NULL,
  sdir = NULL,
  verbose = FALSE,
  question = "<h4>Question</h4>",
  solution = "<h4>Solution</h4>",
  mathjax = NULL,
  resolution = 100,
  width = 4,
  height = 4,
  svg = FALSE,
  encoding = "",
  envir = NULL,
  converter = NULL,
  seed = NULL,
  ...
)
}
\description{
Create a simple Javascript-powered quiz for testing purposes.
}
\details{
#Credits
Modified from exams2html, by Achim.
}
