% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_answer_columns.r
\name{find_answer_columns}
\alias{find_answer_columns}
\title{Find all answer columns in a dataframe.}
\usage{
find_answer_columns(df)
}
\arguments{
\item{df}{Dataframe to search.}
}
\description{
Find all answer columns in a dataframe.
}
\details{
An answer column is any column where the header starts with "ans" or "Ans".
}
\section{Credits}{
Brighton Pauli, 2022.
}

