% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_questions.R
\name{find_questions}
\alias{find_questions}
\title{Search for Rmds with certain metadata.}
\usage{
find_questions(dir, ..., pattern = ".*")
}
\arguments{
\item{dir}{Directory containing exercise .Rmd files to search through.}

\item{...}{Expressions that return a logical value. See \code{filter}.}

\item{pattern}{Optional regular expression to filter file names}
}
\value{
Dataframe with columns \code{filename} and \code{qvariation}.
}
\description{
Locate all exercise files that match the metadata filters.
}
\section{Credits}{
Brighton Pauli, 2023
}

