% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.r
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve all r/exams-specific metadata from a question's .Rmd file.}
\usage{
get_metadata(filename = "", directory = "")
}
\arguments{
\item{filename}{Name of the .Rmd file. Any extensions are ignored.}

\item{directory}{Optional. Folder this file can be found in.}
}
\value{
List of metadata values, named appropriately.
}
\description{
Retrieve all r/exams-specific metadata from a question's .Rmd file.
}
\details{
If \code{filename} is left blank, a file explorer window is opened.
}
\section{Credits}{
Brighton Pauli, 2023
}

