% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img2txt.R
\name{img2txt}
\alias{img2txt}
\title{Convert an image into a very rough text representation.}
\usage{
img2txt(imgpath, h = 10)
}
\arguments{
\item{imgpath}{File path for image.}

\item{h}{Height of result image, in number of lines. (Default 10)}
}
\description{
Convert an image into a very rough text representation.
}
