% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rmd.r
\name{reformat_string_}
\alias{reformat_string_}
\title{Fix strings for use in dynamic R code chunks}
\usage{
reformat_string_(s)
}
\arguments{
\item{s}{String to be reformatted}
}
\value{
String that can be used in a R code
}
\description{
Fix strings for use in dynamic R code chunks
}
