% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_from_pool.r
\name{add_from_pool}
\alias{add_from_pool}
\title{Add a question from a pool of answers.}
\usage{
add_from_pool(
  question,
  image = NA,
  explanation = NA,
  answer_pool = NA,
  correct_ids = NA,
  keywords = NA,
  df = NA
)
}
\arguments{
\item{question}{Question text.}

\item{image}{Image filename (optional).}

\item{explanation}{Explanation for the correct answer (optional).}

\item{answer_pool}{Tibble of answer choices, with id and text columns.}

\item{correct_ids}{Numeric vector of correct answer ids, as in answer_pool.}

\item{df}{Dataframe to add question to.}
}
\value{
The provided dataframe \code{df}, with an additional row. If a dataframe
is not provided, a new one is created.
}
\description{
Add a multiple-choice or single-choice question to a dataframe,
drawing answers from a pool of possibilities and validation and/or
transforming inputs.
}
\section{Credits}{
Brighton Pauli, 2022.
}

\examples{
df <- examling::build_question_df()
df <- add_from_pool(
 question = "What color is a male cardinal?",
 answer_pool = data.frame(text=c("Red", "Green", "Blue"), id = seq_len(3)),
 correct_ids = c(1),
 df = df
)
df

}
\seealso{
\code{add_question()}
}
