% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_question.r
\name{add_question}
\alias{add_question}
\title{Add a question to a dataframe, validating and/or transforming inputs.}
\usage{
add_question(
  question,
  image = NA,
  correct_ids = NA,
  choices = NA,
  correct = NA,
  incorrect = NA,
  explanation = NA,
  df = NA
)
}
\arguments{
\item{question}{Question text.}

\item{image}{Image filename.}

\item{correct_ids}{Numeric vector of correct answer indices.}

\item{choices}{Tibble of answer choices, with id and text columns.}

\item{correct}{Vector of possible correct answers.}

\item{incorrect}{Vector of possible incorrect answers.}

\item{df}{Dataframe to add question to.

If a dataframe is not provided, a new one is created.

ONLY correct_ids+choices OR correct+incorrect should be provided.

If you use the first pair, "correct" and "incorrect" vectors will be
generated automatically from the choices using the given indices.}
}
\description{
Add a question to a dataframe, validating and/or transforming inputs.
}
\section{Credits}{
Brighton Pauli, 2022.
}

