% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_question.r
\name{add_question}
\alias{add_question}
\title{Add a question to a dataframe, validating and/or transforming inputs.}
\usage{
add_question(
  question,
  image = NA,
  explanation = NA,
  correct = NA,
  incorrect = NA,
  keywords = NA,
  df = NA
)
}
\arguments{
\item{question}{Question text.}

\item{image}{Image filename (optional).}

\item{explanation}{Explanation for the correct answer (optional).}

\item{correct}{Vector of possible correct answers.}

\item{incorrect}{Vector of possible incorrect answers.}

\item{df}{Dataframe to add question to.

If you use the first pair, "correct" and "incorrect" vectors will be
generated automatically from the choices using the given indices.}
}
\description{
Add a question to a dataframe, validating and/or transforming inputs.
}
\section{Credits}{
Brighton Pauli, 2022.
}

