% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_string_question.R
\name{add_string_question}
\alias{add_string_question}
\title{Add a string question to a dataframe.}
\usage{
add_string_question(
  question,
  image = NA,
  explanation = NA,
  correct = NA,
  keywords = NA,
  df = NA
)
}
\arguments{
\item{question}{Question text}

\item{image}{Image filename (optional)}

\item{explanation}{Explanation for the correct answer (optional)}

\item{correct}{Correct answer, string}

\item{keywords}{Vector of strings to insert into the question text}

\item{df}{Dataframe to add question to}
}
\value{
The provided dataframe \code{df}, with an additional row. If a dataframe
is not provided, a new one is created.
}
\description{
Add a string-type question to a dataframe, validating and/or
transforming inputs.
}
\section{Credits}{
Brighton Pauli, 2023
}

\seealso{
\code{add_question()}
}
