% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id.R
\name{create_id}
\alias{create_id}
\title{create an ID with format "[category][subcat][random number]"}
\usage{
create_id(category, subcat = NA)
}
\arguments{
\item{category}{The category of the question.}

\item{subcat}{The subcategory of the question (optional).}
}
\value{
A string of the form "[category][subcat][random number]".
}
\description{
create an ID with format "[category][subcat][random number]"
}
\details{
#Credits
Brighton Pauli, 2022.
}
