% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rmd.r
\name{dyna_question_segment_}
\alias{dyna_question_segment_}
\title{Generate a single \code{add_from_pool} call for this row}
\usage{
dyna_question_segment_(index, df, meta_cols)
}
\arguments{
\item{row}{"schoice" or "mchoice" row from a validated rexamsll dataframe}
}
\value{
String that calls \code{add_from_pool}
}
\description{
Generate a single \code{add_from_pool} call for this row
}
\seealso{
\code{add_from_pool.R}
}
