% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rmd.r
\name{dyna_string_question_segment_}
\alias{dyna_string_question_segment_}
\title{Generate a single \code{add_string_question} call for this row}
\usage{
dyna_string_question_segment_(row)
}
\arguments{
\item{row}{"string" row from a validated rexamsll dataframe}
}
\value{
String that calls \code{add_string_question}
}
\description{
Generate a single \code{add_string_question} call for this row
}
\seealso{
\code{add_string_question.R}
}
