% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_func.R
\name{start_logs}
\alias{start_logs}
\title{Wrapper for \code{logr::log_open}.}
\usage{
start_logs(filepath = NA)
}
\arguments{
\item{filepath}{Path to write logs to.}
}
\description{
Create a log file, if needed, and run \code{logr::log_open}.`
}
\details{
By default, the path for logs is "./logs/yyyymmdd_hhmmsss.log"
}
